refactoring di
remove dagger android
optimize dagger modules
change package
