+++ app/build.gradle
@@ -80,12 +80,9 @@ dependencies {
     implementation 'androidx.swiperefreshlayout:swiperefreshlayout:1.1.0'
 
     /*Dagger 2*/
-    def daggerVersion = '2.35.1'
+    def daggerVersion = '2.38.1'
     implementation "com.google.dagger:dagger:$daggerVersion"
-    implementation "com.google.dagger:dagger-android:$daggerVersion"
-    implementation "com.google.dagger:dagger-android-support:$daggerVersion"
     kapt "com.google.dagger:dagger-compiler:$daggerVersion"
-    kapt "com.google.dagger:dagger-android-processor:$daggerVersion"
 
     /*Retrofit*/
     implementation 'com.squareup.retrofit2:retrofit:2.5.0'
