+++ app/src/main/java/com/semisonfire/cloudgallery/App.kt
@@ -1,15 +1,14 @@
 package com.semisonfire.cloudgallery
 
+import android.app.Application
 import android.os.Environment
-import com.semisonfire.cloudgallery.core.di.AppComponent
-import com.semisonfire.cloudgallery.core.di.DaggerAppComponent
+import com.semisonfire.cloudgallery.di.AppComponent
+import com.semisonfire.cloudgallery.di.DaggerAppComponent
 import com.semisonfire.cloudgallery.utils.ExternalFileProvider
 import com.semisonfire.cloudgallery.utils.FileUtils
 import com.squareup.picasso.Picasso
-import dagger.android.AndroidInjector
-import dagger.android.DaggerApplication
 
-class App : DaggerApplication() {
+class App : Application() {
 
     override fun onCreate() {
 
@@ -22,14 +21,9 @@ class App : DaggerApplication() {
         FileUtils.fileProvider = fileProvider
 
         component = DaggerAppComponent.factory().create(this)
-        component.inject(this)
 
         Picasso.setSingletonInstance(component.picasso())
         super.onCreate()
-    }
-
-    override fun applicationInjector(): AndroidInjector<out DaggerApplication> {
-        return component
     }
 
     companion object {
