+++ app/src/main/java/com/semisonfire/cloudgallery/di/annotation/Scopes.kt
@@ -1,0 +1,11 @@
+package com.semisonfire.cloudgallery.di.annotation
+
+import javax.inject.Scope
+
+@Retention(AnnotationRetention.RUNTIME)
+@Scope
+annotation class ActivityScope
+
+@Retention(AnnotationRetention.RUNTIME)
+@Scope
+annotation class FragmentScope
