+++ app/src/main/java/com/semisonfire/cloudgallery/di/builder/ActivityProvider.kt
@@ -1,0 +1,16 @@
+package com.semisonfire.cloudgallery.di.builder
+
+import dagger.Module
+
+@Module
+abstract class ActivityProvider {
+
+//    @ActivityScope
+//    @ContributesAndroidInjector(modules = [MainModule::class])
+//    internal abstract fun contributeMainActivity(): MainActivity
+//
+//    @ActivityScope
+//    @ContributesAndroidInjector(modules = [PhotoDetailModule::class])
+//    internal abstract fun contributePhotoDetailActivity(): PhotoDetailActivity
+}
+
