+++ app/src/main/java/com/semisonfire/cloudgallery/di/builder/FragmentProvider.kt
@@ -1,0 +1,19 @@
+package com.semisonfire.cloudgallery.di.builder
+
+import dagger.Module
+
+@Module
+abstract class FragmentProvider {
+
+//    @FragmentScope
+//    @ContributesAndroidInjector(modules = [DiskModule::class])
+//    internal abstract fun contributeDiskFragment(): DiskFragment
+//
+//    @FragmentScope
+//    @ContributesAndroidInjector(modules = [TrashModule::class])
+//    internal abstract fun contributeTrashFragment(): TrashFragment
+//
+//    @FragmentScope
+//    @ContributesAndroidInjector(modules = [SettingsModule::class])
+//    internal abstract fun contributeSettingsFragment(): SettingsFragment
+}
