+++ app/src/main/java/com/semisonfire/cloudgallery/di/module/ActivityModule.kt
@@ -1,0 +1,9 @@
+package com.semisonfire.cloudgallery.di.module
+
+import com.semisonfire.cloudgallery.di.builder.FragmentProvider
+import dagger.Module
+
+@Module(includes = [FragmentProvider::class])
+abstract class ActivityModule
+
+
