+++ app/src/main/java/com/semisonfire/cloudgallery/ui/disk/di/DiskModule.kt
@@ -1,6 +1,6 @@
 package com.semisonfire.cloudgallery.ui.disk.di
 
-import com.semisonfire.cloudgallery.core.di.FragmentScope
+import com.semisonfire.cloudgallery.di.annotation.FragmentScope
 import com.semisonfire.cloudgallery.ui.disk.DiskPresenter
 import com.semisonfire.cloudgallery.ui.disk.DiskPresenterImpl
 import com.semisonfire.cloudgallery.ui.disk.data.DiskRepository
