+++ app/src/main/java/com/semisonfire/cloudgallery/ui/main/di/MainModule.kt
@@ -2,10 +2,10 @@ package com.semisonfire.cloudgallery.ui.main.di
 
 import androidx.appcompat.app.AppCompatActivity
 import com.semisonfire.cloudgallery.core.data.remote.auth.AuthRepository
-import com.semisonfire.cloudgallery.core.di.ActivityScope
-import com.semisonfire.cloudgallery.core.di.module.ActivityModule
 import com.semisonfire.cloudgallery.core.ui.navigation.Navigator
 import com.semisonfire.cloudgallery.core.ui.navigation.NavigatorImpl
+import com.semisonfire.cloudgallery.di.annotation.ActivityScope
+import com.semisonfire.cloudgallery.di.module.ActivityModule
 import com.semisonfire.cloudgallery.ui.main.MainActivity
 import com.semisonfire.cloudgallery.ui.main.MainPresenter
 import com.semisonfire.cloudgallery.ui.main.MainPresenterImpl
