+++ app/src/main/java/com/semisonfire/cloudgallery/ui/photo/di/PhotoDetailModule.kt
@@ -1,10 +1,10 @@
 package com.semisonfire.cloudgallery.ui.photo.di
 
 import androidx.appcompat.app.AppCompatActivity
-import com.semisonfire.cloudgallery.core.di.ActivityScope
-import com.semisonfire.cloudgallery.core.di.module.ActivityModule
 import com.semisonfire.cloudgallery.core.ui.navigation.Navigator
 import com.semisonfire.cloudgallery.core.ui.navigation.NavigatorImpl
+import com.semisonfire.cloudgallery.di.annotation.ActivityScope
+import com.semisonfire.cloudgallery.di.module.ActivityModule
 import com.semisonfire.cloudgallery.ui.disk.data.DiskRepository
 import com.semisonfire.cloudgallery.ui.photo.PhotoDetailActivity
 import com.semisonfire.cloudgallery.ui.photo.PhotoDetailPresenter
