+++ app/src/main/java/com/semisonfire/cloudgallery/ui/trash/di/TrashModule.kt
@@ -1,6 +1,6 @@
 package com.semisonfire.cloudgallery.ui.trash.di
 
-import com.semisonfire.cloudgallery.core.di.FragmentScope
+import com.semisonfire.cloudgallery.di.annotation.FragmentScope
 import com.semisonfire.cloudgallery.ui.trash.TrashPresenter
 import com.semisonfire.cloudgallery.ui.trash.TrashPresenterImpl
 import com.semisonfire.cloudgallery.ui.trash.data.TrashRepository
